:sparkles: feat: Cliente

+++ TrabalhoFinal/ecommerce/src/main/java/com/serratec/ecommerce/ecommerce/repository/EnderecoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.serratec.ecommerce.ecommerce.model.Endereco;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
 
 }
